+++ Source/Geometry/Square/Square.h
@@ -8,7 +8,7 @@ public:
 	Square(const Point& _p, const Vector& _n, const RGBColor& _c, float _halfSize);
 	~Square();
 
-	bool HitTest(const Ray& _r, float& _t, const float& _tMin) override;
+	bool HitTest(const Ray& _r, double& _t, const float& _tMin) override;
 
 private:
 	float mHalfSize;
